requirements.txt
tqdm
numpy
torch 
torchvision 
torchaudio
sentence-transformers 
transformers
aiofiles
einops
nomic[local]
h5py
faiss-cpu
